fix: resolve PHPStan errors in FHIRPath component
- Add FHIRPath component to PHPStan analysis configuration
- Fix EvaluationException::create() calls to use constructor
- Add null checks for evaluator in function classes
- Cast hexdec() result to int for mb_chr()
- Remove redundant instanceof check in FHIRPathEvaluator
- Remove unnecessary substr() false check
- Add use statements for EvaluationException in function classes
- Generate baseline for remaining lexer/parser static analysis issues

+++ src/Component/FHIRPath/src/Parser/FHIRPathLexer.php
@@ -310,7 +310,7 @@ class FHIRPathLexer
             $hex .= $this->advance();
         }
 
-        return mb_chr(hexdec($hex), 'UTF-8');
+        return mb_chr((int) hexdec($hex), 'UTF-8');
     }
 
     /**
